add support for --reference parameter
